chore: extract selector logic so tests exercise production code
The tests previously duplicated consumeSelector/updateSelectorText as
mocks, which could silently drift from the shipping polyfill. Extract the
pure, DOM-free selector-rewriting logic into scoped-styles-selectors.js,
imported by both the polyfill and the test. The tests now exercise the
exact code that ships.
